# add_edge! and has_edge now allow any type(s) that can be converted to…