… an edge.

I noticed that `add_edge!(g, 1=>3)` didn't work (it used to), so I genericized `add_edge!` and `has_edge` so that they just pass whatever arguments they have to create an edge, and then call the method that uses a constructed edge.

I did *not* do this for `rem_edge!`, though we probably should for consistency.