@jpfairbanks 2017-03-31
We have a small problem @sbromberger. I thought that the varargs was going to be a performance problem, but I didn't want to just write a lame benchmark. So I looked at the x86 code that they produced.

Here we look at the old behavior:
```julia
julia> using LightGraphs

julia> g = Graph(10)
{10, 0} undirected simple Int64 graph

julia> @code_warntype add_edge_old!(g,1,2)
Variables:
  #self#::#add_edge_old!
  g::LightGraphs.SimpleGraphs.SimpleGraph{Int64}
  u::Int64
  v::Int64
  T::Any

Body:
  begin  # line 3:
      return $(Expr(:invoke, MethodInstance for add_edge!(::LightGraphs.SimpleGraphs.SimpleGraph{Int64}, ::LightGraphs.SimpleGraphs.SimpleEdge{Int64}), :(Main.add_edge!), :(g), :($(Expr(:new, :($(QuoteNode(LightGraphs.SimpleGraphs.SimpleEdge{Int64}))), :(u), :(v))))))
  end::Bool

julia> function add_edge_old!(g::AbstractSimpleGraph, u::Integer, v::Integer)                
            T = eltype(g)                
            add_edge!(g, edgetype(g)(T(u),T(v)))                
        end
add_edge_old! (generic function with 1 method)

julia> @code_native add_edge_old!(g, 1,2)
	.text
Filename: REPL[10]
	pushq	%rbp
	movq	%rsp, %rbp
Source line: 3
	subq	$16, %rsp
	movq	%rsi, -16(%rbp)
	movq	%rdx, -8(%rbp)
	movabsq	$"add_edge!", %rax
	leaq	-16(%rbp), %rsi
	callq	*%rax
	addq	$16, %rsp
	popq	%rbp
	retq
	nopw	%cs:(%rax,%rax)

```
Not bad relatively few instructions with only one block --- looks good for a simple function.

Then we look at the new functionality:
```julia
julia> @code_warntype add_edge!(g, 1,2)
Variables:
  #self#::LightGraphs.#add_edge!
  g::LightGraphs.SimpleGraphs.SimpleGraph{Int64}
  x::Tuple{Int64,Int64}

Body:
  begin 
      return $(Expr(:invoke, MethodInstance for add_edge!(::LightGraphs.SimpleGraphs.SimpleGraph{Int64}, ::LightGraphs.SimpleGraphs.SimpleEdge{Int64}), :(LightGraphs.SimpleGraphs.add_edge!), :(g), :($(Expr(:new, :($(QuoteNode(LightGraphs.SimpleGraphs.SimpleEdge{Int64}))), :((Core.getfield)(x, 1)::Int64), :((Core.getfield)(x, 2)::Int64))))))
  end::Bool
```

The Tuple argument in there looks odd, we didn't need any tuples before.
```julia
julia> @code_native add_edge!(g,1,2)
	.text
Filename: SimpleGraphs.jl
	pushq	%rbp
	movq	%rsp, %rbp
	subq	$32, %rsp
	movq	%rsi, %rax
	movq	%rax, -24(%rbp)
Source line: 57
	addl	$-1, %edx
	je	L95
	movq	8(%rax), %rcx
	movq	(%rcx), %rcx
	movq	%rcx, -16(%rbp)
	cmpl	$1, %edx
	je	L121
	movq	(%rax), %rdi
	movq	16(%rax), %rax
	movq	(%rax), %rax
	movq	%rax, -8(%rbp)
	movabsq	$"add_edge!", %rax
	leaq	-16(%rbp), %rsi
	callq	*%rax
	movabsq	$139885862146144, %rdx  # imm = 0x7F39B71F5060
	leaq	16(%rdx), %rcx
	testb	$1, %al
	cmoveq	%rdx, %rcx
	movq	%rcx, %rax
	addq	$32, %rsp
	popq	%rbp
	retq
L95:
	addq	$8, %rax
	movabsq	$jl_bounds_error_tuple_int, %rcx
	xorl	%esi, %esi
	movl	$1, %edx
	movq	%rax, %rdi
	callq	*%rcx
L121:
	addq	$8, %rax
	movabsq	$jl_bounds_error_tuple_int, %rcx
	movl	$1, %esi
	movl	$2, %edx
	movq	%rax, %rdi
	callq	*%rcx
```

That is many more instructions! Including 2 blocks for handling errors. 
We can add a benchmark for insertion times to measure the performance penalty.