@sbromberger 2017-03-31
Yeah, that appears to be a big problem. Perhaps we just fix it by writing explicit converters - that is, a one-argument converter (for tuple and pair) and then a two-argument converter (for src,dst)?