@sbromberger 2017-04-02
current code benchmark:
```
julia> @benchmark  g = Graph(100000, 16*100000)
BenchmarkTools.Trial:
  memory estimate:  339.69 MiB
  allocs estimate:  3756669
  --------------
  minimum time:     1.648 s (13.87% GC)
  median time:      1.810 s (21.97% GC)
  mean time:        1.824 s (20.93% GC)
  maximum time:     2.013 s (25.78% GC)
  --------------
  samples:          3
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

this PR benchmark:
```
julia> @benchmark  g = Graph(100000, 16*100000)
BenchmarkTools.Trial:
  memory estimate:  339.71 MiB
  allocs estimate:  3756660
  --------------
  minimum time:     1.602 s (13.64% GC)
  median time:      1.811 s (21.30% GC)
  mean time:        1.749 s (20.01% GC)
  maximum time:     1.834 s (24.29% GC)
  --------------
  samples:          3
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```