@sbromberger 2017-04-02
@jpfairbanks - this new code could use your review :) Thanks.