@jpfairbanks 2017-04-03
I the idea of posting the code native. Everyone should look at some assembly every so often :)