@sbromberger 2017-04-03
> Everyone should look at some assembly every so often 

... to remind them to thank their deity that they no longer have to deal with it? :)