@jpfairbanks 2017-04-07
Yes! and to make sure that they remember that all PLs eventually turn into instructions.